Revert "refactor: clean up the projectPropContext"
This reverts commit 5ba0b428d72494f0003e2b495e55b31b0d076782.
